requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
blinker==1.8.2
boto3==1.34.63
botocore==1.34.63
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.2.1
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.0
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
distlib==0.3.7
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-heroku==0.3.1
django-model-utils==4.4.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flake8==7.0.0
flake8-bugbear==24.4.26
flake8-comprehensions==3.14.0
Flask==3.0.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gunicorn==21.2.0
h11==0.14.0
honcho==1.1.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.1
importlib_metadata==7.1.0
incremental==22.10.0
intel-openmp==2021.4.0
iopath==0.1.10
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.4
libcst==1.3.1
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mkl==2021.4.0
moreorless==0.4.0
mpmath==1.2.1
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
open3d==0.18.0
openai==1.37.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
phonenumberslite==8.13.31
pillow==10.3.0
platformdirs==4.0.0
plotly==5.22.0
portalocker==2.8.2
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypng==0.20220715.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.18.1
s3transfer==0.10.1
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
service-identity==24.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
stdlibs==2024.1.28
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tenacity==8.3.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.5.3
toml==0.10.2
tomlkit==0.12.5
torch==2.3.0
torchaudio==2.2.0.dev20240507+cu118
torchvision==0.18.0
tqdm==4.66.4
trailrunner==1.4.0
traitlets==5.14.3
twilio==9.0.0
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
usort==1.0.8.post1
vine==5.1.0
virtualenv==20.24.7
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
widgetsnbextension==4.0.10
yacs==0.1.8
yarl==1.9.4
zipp==3.19.0
zope.interface==6.2
